requirements.txt
vk.api == 11.9.4
vk == 2.0.2
python-dotenv == 0.20.0